refactor: systems

+++ lc_client/src/lc_client/eng_graphics/openGL/gl_shader_work.cpp
@@ -8,7 +8,7 @@
 void ShaderWorkGl::loadShaders(
 	entt::entity entity, const std::string vertexShaderName, const std::string fragmentShaderName) {
 
-	m_pRegistry->emplace_or_replace<ShaderGl>(entity, createShaderProgram(vertexShaderName, fragmentShaderName));
+	m_pRegistry->emplace_or_replace<ShaderGl>(entity, (int) createShaderProgram(vertexShaderName, fragmentShaderName));
 }
 
 unsigned int ShaderWorkGl::createShaderProgram(std::string vertexShaderName, std::string fragmentShaderName) {
